requirement.txt
lmdeploy
python-pptx
openai
gradio
timm
'pybind11>=2.12'.